Create README.md
Following the installation of the header, the program will prompt the user to choose the type of gas and specify the number of gallons they wish to pump. Once the necessary information is obtained, the gas price will be multiplied by the gallons requested to determine the total cost for refueling. Next, the total value will be calculated by adding the sales tax to the calculated gas price. Subsequently, the user will be asked to select a payment method for completing the transaction, such as credit card, debit card, gift card, or cash.