Create Assignment 1 - Gas Station
After implementing the header, the software prompts the user to select the gas type and amount of gallons to fill. To calculate the total cost of filling up the gas tank, multiply the gas price by the requested number of gallons. To get the total amount, multiply the gas price per gallon by the sales tax. Ask the user for their preferred payment option, such as credit card, debit card, gift card, or cash.