Create Assignment 1 - Country Club Menu
This platform allows you to order meals. This program displays the menu and prompts the user to select their desired meal and dessert. After adding the dinner and dessert, compute the subtotal and sales tax. After adding the subtotal and sales tax, present the total figure to the user before terminating the program.